Fixed compass bug. Fetch data in a thread. Added vertical speed indicator

+++ src/com/example/testvaadin/javascriptcomponents/PrimaryFlightDisplayState.java
@@ -5,12 +5,12 @@ import com.vaadin.shared.ui.JavaScriptComponentState;
 @SuppressWarnings("serial")
 public class PrimaryFlightDisplayState extends JavaScriptComponentState {
 	public String xhtml;
-	public int altitude;
+	public float altitude;
 	public int speed;
 	public int roll;
 	public int pitch;
 	public int heading;
 	public int truecourse;
 	public int resetpfd;
-	public int verticalspeed;
+	public float verticalspeed;
 }
